Reduce wastage in debug CPU cycles
Ultra laudatory change:
- Removed a wasteful double assignment and calculation for the size
  of f->iosz. This error appears to have been introduced in
  2003-06-21 ksh93o+, and likely would have gone unnoticed if not
  for the sign conversion. (In practice this line was optimized out
  by the compiler regardless.)

+++ src/lib/libast/disc/sfdcseekable.c
@@ -44,7 +44,7 @@ static ssize_t skwrite(Sfio_t*		f,	/* stream involved */
 	NOT_USED(buf);
 	NOT_USED(n);
 	NOT_USED(disc);
-	return (ssize_t)(-1);
+	return -1;
 }
 
 static ssize_t skread(Sfio_t*	f,	/* stream involved */
